Move TP tests to TP directories
