+++ consensus/poet/families/tests/test_tp_validator_registry.py
@@ -19,7 +19,7 @@ import hashlib
 
 from sawtooth_processor_test.transaction_processor_test_case \
     import TransactionProcessorTestCase
-from sawtooth_integration.message_factories.validator_reg_message_factory \
+from sawtooth_validator_registry.validator_reg_message_factory \
     import ValidatorRegistryMessageFactory
 
 from sawtooth_poet_common import sgx_structs
