+++ integration/sawtooth_integration/docker/intkey-javascript-smoke.yaml
@@ -83,5 +83,6 @@ services:
     stop_signal: SIGKILL
     environment:
       PYTHONPATH: "/project/sawtooth-core/sdk/python:\
+        /project/sawtooth-core/sdk/examples/intkey_python:\
         /project/sawtooth-core/integration:\
         /project/sawtooth-core/signing"
