+++ integration/sawtooth_integration/docker/poet-smoke.yaml
@@ -223,7 +223,7 @@ services:
     stop_signal: SIGKILL
     environment:
       PYTHONPATH: "/project/sawtooth-core/sdk/python:\
-        /project/sawtooth-core/integration:\
+        /project/sawtooth-core/sdk/examples/intkey_python:\
         /project/sawtooth-core/signing:\
         /project/sawtooth-core/cli"
