+++ integration/sawtooth_integration/docker/two_families.yaml
@@ -95,5 +95,6 @@ services:
     stop_signal: SIGKILL
     environment:
       PYTHONPATH: "/project/sawtooth-core/sdk/python:\
+        /project/sawtooth-core/sdk/examples/intkey_python:\
         /project/sawtooth-core/integration:\
         /project/sawtooth-core/signing"
