+++ integration/sawtooth_integration/tests/test_intkey_smoke.py
@@ -25,8 +25,7 @@ from base64 import b64decode
 
 import cbor
 
-from sawtooth_integration.message_factories.intkey_message_factory \
-    import IntkeyMessageFactory
+from sawtooth_intkey.intkey_message_factory import IntkeyMessageFactory
 
 
 LOGGER = logging.getLogger(__name__)
