+++ sdk/examples/intkey_python/tests/test_tp_intkey.py
@@ -17,8 +17,7 @@ import cbor
 
 from sawtooth_processor_test.transaction_processor_test_case \
     import TransactionProcessorTestCase
-from sawtooth_integration.message_factories.intkey_message_factory \
-    import IntkeyMessageFactory
+from sawtooth_intkey.intkey_message_factory import IntkeyMessageFactory
 from sawtooth_sdk.protobuf.validator_pb2 import Message
