+++ sdk/examples/intkey_python/tests/tp-intkey-java.yaml
@@ -17,13 +17,13 @@ version: "2.1"
 
 services:
 
-  tp_intkey_python:
-    image: sawtooth-tp_intkey_python:$ISOLATION_ID
+  tp_intkey_java:
+    image: sawtooth-tp_intkey_java:$ISOLATION_ID
     volumes:
       - $SAWTOOTH_CORE:/project/sawtooth-core
     expose:
       - 40000
-    command: tp_intkey_python -vv tcp://validator:40000
+    command: tp_intkey_java tcp://validator:40000
     stop_signal: SIGKILL
 
   validator:
@@ -33,10 +33,11 @@ services:
     expose: 
       - 40000
     command: nose2-3 -v -s
-      /project/sawtooth-core/integration/sawtooth_integration/tests
+      /project/sawtooth-core/sdk/examples/intkey_python/tests
       test_tp_intkey
     stop_signal: SIGKILL
     environment:
       PYTHONPATH: "/project/sawtooth-core/sdk/python:\
+        /project/sawtooth-core/sdk/examples/intkey_python:\
         /project/sawtooth-core/integration:\
         /project/sawtooth-core/signing"
