+++ sdk/examples/xo_python/sawtooth_xo/xo_message_factory.py
