+++ sdk/examples/xo_python/tests/tp-xo-python.yaml
@@ -17,13 +17,13 @@ version: "2.1"
 
 services:
 
-  tp_intkey_javascript:
-    image: sawtooth-tp_intkey_javascript:$ISOLATION_ID
+  tp_xo_python:
+    image: sawtooth-tp_xo_python:$ISOLATION_ID
     volumes:
       - $SAWTOOTH_CORE:/project/sawtooth-core
     expose:
       - 40000
-    command: tp_intkey_javascript tcp://validator:40000
+    command: tp_xo_python -vv tcp://validator:40000
     stop_signal: SIGKILL
 
   validator:
@@ -33,10 +33,12 @@ services:
     expose: 
       - 40000
     command: nose2-3 -v -s
-      /project/sawtooth-core/integration/sawtooth_integration/tests
-      test_tp_intkey
+      /project/sawtooth-core/sdk/examples/xo_python/tests
+      test_tp_xo
     stop_signal: SIGKILL
     environment:
-      PYTHONPATH: "/project/sawtooth-core/sdk/python:\
-        /project/sawtooth-core/integration:\
-        /project/sawtooth-core/signing"
+      PYTHONPATH: "\
+        /project/sawtooth-core/sdk/python:\
+        /project/sawtooth-core/sdk/examples/xo_python:\
+        /project/sawtooth-core/signing:\
+      "
